# Fix network.traceroute exec module function