### What does this PR do?
Armoring the unprotected list dereferences and redirecting the problematic lines to the log seemed like the best solution to this optimistically quixotic function.  Perhaps a future enterprising adventurer can dig into the source code of the many variants of traceroute and with unit tests provide a more orthogonal result.

### Previous Behavior
```py
# salt-call --local -l warning network.traceroute gentoo.org
[ERROR   ] An un-handled exception was caught by salt's global exception handler:
IndexError: list index out of range
Traceback (most recent call last):
  File "/usr/lib/python-exec/python2.7/salt-call", line 11, in <module>
    salt_call()
  File "/usr/lib/python2.7/site-packages/salt/scripts.py", line 431, in salt_call
    client.run()
  File "/usr/lib/python2.7/site-packages/salt/cli/call.py", line 57, in run
    caller.run()
  File "/usr/lib/python2.7/site-packages/salt/cli/caller.py", line 138, in run
    ret = self.call()
  File "/usr/lib/python2.7/site-packages/salt/cli/caller.py", line 237, in call
    ret['return'] = self.minion.executors[fname](self.opts, data, func, args, kwargs)
  File "/usr/lib/python2.7/site-packages/salt/executors/direct_call.py", line 12, in execute
    return func(*args, **kwargs)
  File "/usr/lib/python2.7/site-packages/salt/modules/network.py", line 941, in traceroute
    'ms3': float(comps[4].split()[0])}
IndexError: list index out of range
```
### New Behavior
```yaml
# salt-call --local -l warning network.traceroute gentoo.org
[WARNING ] Cannot parse traceroute output line:  8  be3037.ccr21.den01.atlas.cogentco.com (154.54.41.146)  49.349 ms be3038.ccr22.den01.atlas.cogentco.com (154.54.42.98)  50.383 ms *
[WARNING ] Cannot parse traceroute output line: 15  be3579.rcr51.lba01.atlas.cogentco.com (154.54.57.250)  162.295 ms *  162.492 ms
local:
    [...]
    |_
      ----------
      count:
          19
      hostname:
          www.gentoo.org
      ip: 
          89.16.167.134
      ms1:
          168.193
      ms2:
          163.027
      ms3:
          163.405
```
### What issues does this PR fix or reference?
#21927

### Tests written?
No

### Commits signed with GPG?
Yes